Organize the project and remove home-assistant-isms
